Renamed command
-Renamed "currentlogonsession" to "logonsession"
-Added enumeration of all logonsession information if elevated
