+++ Rubeus/Domain/CommandCollection.cs
@@ -23,7 +23,7 @@ namespace Rubeus.Domain
             _availableCommands.Add(Changepw.CommandName, () => new Changepw());
             _availableCommands.Add(Createnetonly.CommandName, () => new Createnetonly());
             _availableCommands.Add(Currentluid.CommandName, () => new Currentluid());
-            _availableCommands.Add(Currentlogonsession.CommandName, () => new Currentlogonsession());
+            _availableCommands.Add(Logonsession.CommandName, () => new Logonsession());
             _availableCommands.Add(Describe.CommandName, () => new Describe());
             _availableCommands.Add(Dump.CommandName, () => new Dump());
             _availableCommands.Add(Hash.CommandName, () => new Hash());
