+++ Rubeus/Rubeus.csproj
@@ -80,7 +80,7 @@
     <Compile Include="Commands\Brute.cs" />
     <Compile Include="Commands\Changepw.cs" />
     <Compile Include="Commands\Createnetonly.cs" />
-    <Compile Include="Commands\Currentlogonsession.cs" />
+    <Compile Include="Commands\Logonsession.cs" />
     <Compile Include="Commands\Currentluid.cs" />
     <Compile Include="Commands\Describe.cs" />
     <Compile Include="Commands\Dump.cs" />
